+++ example/example.dart
@@ -1,0 +1,6 @@
+import 'package:named_mode/named_mode.dart';
+
+void main() {
+  String s = 'ACore_appleOreo_BabiQoo';
+  print(autoSplit(s, recursive: true));
+}