+++ CHANGELOG.md
@@ -9,3 +9,9 @@
 ## 1.0.0+2
 
 - fixed some bugs
+
+## 1.0.1
+
+- new method `List<String> autoSplit(String src, {NamedMode parentNamedMode = NamedMode.unknown, bool recursive = true})`
+
+- optimize regex pattern